pdf-to-audio-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pdfplumber==0.11.4
PyMuPDF==1.24.14
httpx==0.28.1
python-dotenv==1.0.1
aiofiles==24.2.0
